amount input validation

+++ database/migrations/2024_01_30_121855_create_credits_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
                 ->constrained()
                 ->onUpdate('cascade')
                 ->onDelete('cascade');
-            $table->smallInteger('type');
+            $table->unsignedTinyInteger('type');
             $table->float('amount', 8, 2);
             $table->timestamps();
         });
